# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AL-AQSA Camp API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/LoginInput" } } } }
      responses:
        "200": { description: Authenticated, content: { application/json: { schema: { $ref: "#/components/schemas/SessionResponse" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /auth/session:
    get:
      operationId: getSession
      tags: [auth]
      responses:
        "200": { description: Current session, content: { application/json: { schema: { $ref: "#/components/schemas/SessionResponse" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      responses:
        "200": { description: Logged out, content: { application/json: { schema: { $ref: "#/components/schemas/ActionResponse" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /dashboard:
    get:
      operationId: getDashboard
      tags: [dashboard]
      responses:
        "200": { description: Dashboard summary, content: { application/json: { schema: { $ref: "#/components/schemas/DashboardSummary" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /delegates:
    get:
      operationId: listDelegates
      tags: [delegates]
      parameters: [{ $ref: "#/components/parameters/Page" }, { $ref: "#/components/parameters/Limit" }]
      responses:
        "200": { description: Delegates, content: { application/json: { schema: { $ref: "#/components/schemas/DelegateList" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      operationId: createDelegate
      tags: [delegates]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/DelegateInput" } } } }
      responses:
        "201": { description: Delegate created, content: { application/json: { schema: { $ref: "#/components/schemas/Delegate" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /delegates/{id}:
    get:
      operationId: getDelegate
      tags: [delegates]
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        "200": { description: Delegate, content: { application/json: { schema: { $ref: "#/components/schemas/Delegate" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      operationId: updateDelegate
      tags: [delegates]
      parameters: [{ $ref: "#/components/parameters/Id" }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/DelegateUpdate" } } } }
      responses:
        "200": { description: Delegate updated, content: { application/json: { schema: { $ref: "#/components/schemas/Delegate" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /displaced-persons:
    get:
      operationId: listDisplacedPersons
      tags: [displaced]
      parameters: [{ $ref: "#/components/parameters/Page" }, { $ref: "#/components/parameters/Limit" }]
      responses:
        "200": { description: Displaced persons, content: { application/json: { schema: { $ref: "#/components/schemas/DisplacedPersonList" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      operationId: createDisplacedPerson
      tags: [displaced]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/DisplacedPersonInput" } } } }
      responses:
        "201": { description: Person created, content: { application/json: { schema: { $ref: "#/components/schemas/DisplacedPerson" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /displaced-persons/{id}:
    get:
      operationId: getDisplacedPerson
      tags: [displaced]
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        "200": { description: Person, content: { application/json: { schema: { $ref: "#/components/schemas/DisplacedPerson" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      operationId: updateDisplacedPerson
      tags: [displaced]
      parameters: [{ $ref: "#/components/parameters/Id" }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/DisplacedPersonUpdate" } } } }
      responses:
        "200": { description: Person updated, content: { application/json: { schema: { $ref: "#/components/schemas/DisplacedPerson" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /displaced-persons/{id}/delegate:
    patch:
      operationId: assignDisplacedPersonDelegate
      tags: [displaced]
      parameters: [{ $ref: "#/components/parameters/Id" }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/DelegateAssignmentInput" } } } }
      responses:
        "200": { description: Delegate assigned, content: { application/json: { schema: { $ref: "#/components/schemas/DisplacedPerson" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /aids:
    get:
      operationId: listAids
      tags: [aids]
      parameters: [{ $ref: "#/components/parameters/Page" }, { $ref: "#/components/parameters/Limit" }]
      responses:
        "200": { description: Aid campaigns, content: { application/json: { schema: { $ref: "#/components/schemas/AidList" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      operationId: createAid
      tags: [aids]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/AidInput" } } } }
      responses:
        "201": { description: Aid created, content: { application/json: { schema: { $ref: "#/components/schemas/Aid" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /aids/{id}:
    get:
      operationId: getAid
      tags: [aids]
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        "200": { description: Aid campaign, content: { application/json: { schema: { $ref: "#/components/schemas/Aid" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      operationId: updateAid
      tags: [aids]
      parameters: [{ $ref: "#/components/parameters/Id" }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/AidUpdate" } } } }
      responses:
        "200": { description: Aid updated, content: { application/json: { schema: { $ref: "#/components/schemas/Aid" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /aids/{id}/recipients:
    post:
      operationId: addAidRecipients
      tags: [aids]
      parameters: [{ $ref: "#/components/parameters/Id" }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/AidRecipientsInput" } } } }
      responses:
        "200": { description: Recipients added, content: { application/json: { schema: { $ref: "#/components/schemas/Aid" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /aids/{id}/receive:
    post:
      operationId: receiveAid
      tags: [aids]
      parameters: [{ $ref: "#/components/parameters/Id" }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/AidReceiptInput" } } } }
      responses:
        "200": { description: Receipt recorded, content: { application/json: { schema: { $ref: "#/components/schemas/ActionResponse" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /complaints:
    get:
      operationId: listComplaints
      tags: [complaints]
      parameters: [{ $ref: "#/components/parameters/Page" }, { $ref: "#/components/parameters/Limit" }]
      responses:
        "200": { description: Complaints, content: { application/json: { schema: { $ref: "#/components/schemas/ComplaintList" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      operationId: createComplaint
      tags: [complaints]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/ComplaintInput" } } } }
      responses:
        "201": { description: Complaint created, content: { application/json: { schema: { $ref: "#/components/schemas/Complaint" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      parameters: [{ $ref: "#/components/parameters/Page" }, { $ref: "#/components/parameters/Limit" }]
      responses:
        "200": { description: Notifications, content: { application/json: { schema: { $ref: "#/components/schemas/NotificationList" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
components:
  parameters:
...
    Page: { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
    Limit: { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
  responses:
    BadRequest: { description: Invalid request, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
    Unauthorized: { description: Authentication required, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
    Forbidden: { description: Insufficient permissions, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorResponse" } } } }
...
  schemas:
...
    SessionResponse:
      type: object
      required: [user, token, expiresAt]
      properties:
        user: { $ref: "#/components/schemas/User" }
        token: { type: string }
        expiresAt: { type: string, format: date-time }
    DashboardSummary:
      type: object
      required: [totalDelegates, totalDisplacedPersons, activeAids, openComplaints, unreadNotifications]
      properties:
        totalDelegates: { type: integer }
        totalDisplacedPersons: { type: integer }
        activeAids: { type: integer }
        openComplaints: { type: integer }
        unreadNotifications: { type: integer }
...
    ActionResponse:
      type: object
      required: [success, message]
      properties: { success: { type: boolean }, message: { type: string } }
...
    LoginInput:
      type: object
      required: [email, password, role]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 1 }
        role: { type: string, enum: [admin, manager, delegate, displaced, security] }
...
    DelegateList: { $ref: "#/components/schemas/DelegatePage" }
...
    AidList: { $ref: "#/components/schemas/AidPage" }
...